% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeLineGraphics.R
\name{aspectRatio}
\alias{aspectRatio}
\title{aspectRatio}
\usage{
aspectRatio(rim)
}
\arguments{
\item{rim}{is rasterized version of im}
}
\value{
numerical ratio needed for maintaining proper aspect ratio in graph overlay image
}
\description{
compute ratio needed for maintaining proper aspect ratio in graph overlay image
}
\examples{
aspectRatio(as.raster(magick::image_read(example2()[[1]]$pics[[1]])))
}
