% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeLineGraphics.R
\name{backFitRaster}
\alias{backFitRaster}
\title{backFitRaster}
\usage{
backFitRaster(y2x, maxy, ymargin)
}
\arguments{
\item{y2x}{is original aspect ratio of image}

\item{maxy}{is max vertical size in inches}

\item{ymargin}{is fraction of maxy to leave as top and bottom margin}
}
\value{
numerical x size of raster image
}
\description{
given a maximum vertical size in inches that is allowable, find x size of raster image
}
\details{
graph units<-yinch(y in inches)
}
\examples{
backFitRaster(1.333,10,.05)
}
