% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeLineGraphics.R
\name{nextMultiple10}
\alias{nextMultiple10}
\title{nextMultiple10}
\usage{
nextMultiple10(x)
}
\arguments{
\item{x}{integer}
}
\value{
integer next higher multiple of 10
}
\description{
find next higher multiple of 10
}
\examples{
nextMultiple10(1732)
nextMultiple10(1699)
}
