% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeLineGraphics.R
\name{roy}
\alias{roy}
\title{roy}
\usage{
roy(l)
}
\arguments{
\item{l}{is a list
\itemize{
\item character string $name the person's name
\item integer $start starting year
\item integer $end ending year
\item character string $color background color for the horizontal line
\item list of overlay pictures $pics, each entry contains
\itemize{
\item character string path name for a picture
\item integer year for placement of left edge of picture
}
}}
}
\value{
integer vector containing the range of years
}
\description{
determine the range of years
}
\examples{
roy(example2())
}
