% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeLineGraphics.R
\name{stripPlot}
\alias{stripPlot}
\title{stripPlot}
\usage{
stripPlot(l, ht, vfactor, ymargin, nameCEX, leftOver)
}
\arguments{
\item{l}{is a list
\itemize{
\item character string $name the person's name
\item integer $start starting year
\item integer $end ending year
\item character string $color background color for the horizontal line
\item list of overlay pictures $pics, each entry contains
\itemize{
\item character string path name for a picture
\item integer year for placement of left edge of picture
}
}}

\item{ht}{is total height of entire graph}

\item{vfactor}{is height of a horizontal strip}

\item{ymargin}{is fraction of maxy to leave as top and bottom margin}

\item{nameCEX}{control size of font in text() to display names associated with each horizontal strip}

\item{leftOver}{if TRUE then pictures that do not fit in horizontal strip are placed to left of strip rather than omitted}
}
\value{
{No return value, called for side effects}
}
\description{
iteratively plot each horizontal strip
}
